:mailbox: Reduce batch size and increase retry count
Done in order to try and reduce the likelyhood the requests will be seen
as a DOS attack which ultimately means that some of the download
requests fail. Increasing the number of retry attempts to try and
minimise any requests that are blocked will be able to get through
within the 4 other attempts.